add median class between PUMS aggregator and base pums demographic medians
